[generator] Move the attributes to their own file.
An IKVM-based generator will have the attributes in a separate assembly,
so they need to be in a different file.
