+++ src/Makefile.generator
@@ -3,6 +3,7 @@ GENERATOR_SOURCES =                                             \
 	$(TOP)/src/btouch.cs                               \
 	$(TOP)/src/error.cs                                \
 	$(TOP)/src/generator.cs                            \
+	$(TOP)/src/generator-attributes.cs                 \
 	$(TOP)/src/generator-enums.cs                      \
 	$(TOP)/src/generator-filters.cs                    \
 	$(TOP)/src/ObjCRuntime/Stret.cs                    \
